Rename clipped_factorization_analysis.py to clipped_factorization.py
